requirements.txt
streamlit
google-generativeai
matplotlib
pandas